Extracted a StepParser class
